requirements.txt
django
django-filer
django-mptt
django-polymorphic
easy_thumbnails
psycopg2
Pillow
uwsgi

